refactor(openclaw): extract query result formatters
